% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePresentation.R
\name{GeneratePresentation}
\alias{GeneratePresentation}
\title{Generate presentation}
\usage{
GeneratePresentation(centre.name, this.meeting.date,
  last.meeting.date = NULL, data.path, codebook.path,
  codebook.file.name = "codebook.csv")
}
\arguments{
\item{centre.name}{Character vector of length 1. The name of the centre. No
default.}

\item{this.meeting.date}{Character vector of length 1. The date whan the
meeting is. Should be formatted as YYYY-MM-DD. No default.}

\item{last.meeting.date}{Character vector of length 1. The date whan the last
meeting was or NULL, in which case it is assumed that no previous
meetings have been held and all data is used. Should be formatted as
YYYY-MM-DD. Default is NULL.}

\item{data.path}{Character vector of length 1. The path to the raw data
directory. No default.}

\item{codebook.path}{Character vector of length 1. The path to the data
codebook. No default.}

\item{codebook.file.name}{Character vector of length 1. The codebook file
name. Defaults to "codebook.csv".}
}
\description{
Main function for generating the audit filter review report.
}
